Fix: pass failure details from system runner
