+++ internal/testrunner/runners/pipeline/runner.go
@@ -18,6 +18,7 @@ import (
 	"github.com/elastic/elastic-package/internal/multierror"
 	"github.com/elastic/elastic-package/internal/packages"
 	"github.com/elastic/elastic-package/internal/testrunner"
+	"github.com/elastic/elastic-package/internal/testrunner/runners/testerrors"
 )
 
 const (
@@ -93,7 +94,7 @@ func (r *runner) run() ([]testrunner.TestResult, error) {
 
 		tr.TimeElapsed = time.Now().Sub(startTime)
 		err = r.verifyResults(testCaseFile, result, fieldsValidator)
-		if e, ok := err.(ErrTestCaseFailed); ok {
+		if e, ok := err.(testerrors.ErrTestCaseFailed); ok {
 			tr.FailureMsg = e.Error()
 			tr.FailureDetails = e.Details
 
@@ -169,7 +170,7 @@ func (r *runner) verifyResults(testCaseFile string, result *testResult, fieldsVa
 	}
 
 	err := compareResults(testCasePath, result)
-	if _, ok := err.(ErrTestCaseFailed); ok {
+	if _, ok := err.(testerrors.ErrTestCaseFailed); ok {
 		return err
 	}
 	if err != nil {
@@ -193,7 +194,7 @@ func verifyFieldsInTestResult(result *testResult, fieldsValidator *fields.Valida
 	}
 
 	if len(multiErr) > 0 {
-		return ErrTestCaseFailed{
+		return testerrors.ErrTestCaseFailed{
 			Reason:  "one or more problems with fields found in document",
 			Details: multiErr.Error(),
 		}
